+++ packages/react-relay/vitest.config.ts
@@ -1,14 +1,13 @@
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    globals: true,
-    environment: 'jsdom',
+    environment: "jsdom",
   },
   resolve: {
     alias: {
-      '@bloqz/relay': new URL(
-        '../relay/src/index.ts',
+      "@bloqz/relay": new URL(
+        "../relay/src/index.ts",
         import.meta.url
       ).pathname,
     },