+++ packages/react/vitest.config.ts
@@ -1,9 +1,8 @@
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    globals: true,
-    environment: 'jsdom',
-    jsx: 'react-jsx',
+    environment: "jsdom",
+    jsx: "react-jsx",
   },
 });