+++ packages/concurrency/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2026-08-13
+
+### Changed
+- Updated the README and JSDoc examples to use the `handlers` object API, replacing stale `bloc.on` examples.
+
 ## [1.2.0] - 2025-12-03
 
 ### Changed